implement schema generator
