+++ gql_code_gen/build.yaml
@@ -41,3 +41,9 @@ builders:
     build_extensions: {".graphql": [".input.gql.dart"]}
     auto_apply: none
     build_to: source
+  schema_builder:
+    import: "package:gql_code_gen/schema_builder.dart"
+    builder_factories: ["schemaBuilder"]
+    build_extensions: {".graphql": [".schema.gql.dart"]}
+    auto_apply: none
+    build_to: source
